# swaggerOptions.yml
openapi: 3.0.0
info:
  title: "RespirAR API"
  version: "1.0.0"
  description: "RespirAR API Documentation"
servers:
  - url: "http://localhost:8080/"
    description: "Local server"
paths:    
...
  /api/v1/stations:
    get:
      summary: Obtains stations.
      security:
        - bearerAuth: []
      description: Obtains stations according to parameters.
      tags:
        - Stations
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
          description: Access token for authentication.
        - in: query
          name: pageSize
          schema:
            type: integer
          description: The number of items to return per page.
        - in: query
          name: page
          schema:
            type: integer
          description: The page number to return.
        - in: query
          name: name
          schema:
            type: string
          description: The station name to filter by.
        - in: query
          name: serialNumber
          schema:
            type: string
          description: The serial number to filter by.
        - in: query
          name: status
          schema:
            type: string
          description: The request status to filter by.
        - in: query
          name: date
          schema:
            type: string
          description: The creation date to filter by.
      responses:
        200:
          description: A list of stations with optional pagination.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: Unique identifier of the request.
                      example: "64612f41b0a9b6ebef6a0e66"
                    serial_number:
                      type: string
                      description: Serial number of the station.
                      example: "8edsf23489"
                    name:
                      type: string
                      description: Name of the station.
                      example: "Air-Quality-Station"
                    longitude:
                      type: number
                      format: float
                      description: Longitude of the station's location.
                      example: -159.6969448413211
                    latitude:
                      type: number
                      format: float
                      description: Latitude of the station's location.
                      example: 21.371452961552126
                    brand:
                      type: string
                      description: Brand of the station.
                      example: Huawei
                    model:
                      type: string
                      description: Model of the station.
                      example: L17GYT-G
                    status:
                      type: string
                      enum:
                        - ACTIVE
                        - INACTIVE
                      description: Status of the request.
                    created_by:
                      type: integer
                      description: Identifier of the user who created the request.
                      example: 2
                    created_at:
                      type: string
                      format: date-time
                      description: Date and time when the request was created.
        401:
          description: Unauthorized Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Unathorized"
        403:
          description: Forbbiden Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Forbbiden"
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Internal Server Error"
...
  /api/v1/users/{userId}/stations:
    get:
      summary: Get user stations
      security:
        - bearerAuth: []
      tags:
        - Stations
      description: Retrieve stations for a specific user
      parameters:
        - name: userId
          in: path
          description: ID of the user
          required: true
          schema:
            type: string
        - in: header
          name: Authorization
          schema:
            type: string
          required: true
          description: Access token for authentication.
        - name: pageSize
          in: query
          description: Number of stations to return per page
          schema:
            type: integer
        - name: page
          in: query
          description: Page number of stations to retrieve
          schema:
            type: integer
        - name: name
          in: query
          description: Filter stations by name
          schema:
            type: string
        - name: serialNumber
          in: query
          description: Filter stations by serial number
          schema:
            type: string
        - name: status
          in: query
          description: Filter stations by status
          schema:
            type: string
        - name: date
          in: query
          description: Filter stations by date
          schema:
            type: string
      responses:
        '200':
          description: A list of stations with optional pagination.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: Unique identifier of the request.
                      example: "64612f41b0a9b6ebef6a0e66"
                    serial_number:
                      type: string
                      description: Serial number of the station.
                      example: "8edsf23489"
                    name:
                      type: string
                      description: Name of the station.
                      example: "Air-Quality-Station"
                    longitude:
                      type: number
                      format: float
                      description: Longitude of the station's location.
                      example: -159.6969448413211
                    latitude:
                      type: number
                      format: float
                      description: Latitude of the station's location.
                      example: 21.371452961552126
                    brand:
                      type: string
                      description: Brand of the station.
                      example: Huawei
                    model:
                      type: string
                      description: Model of the station.
                      example: L17GYT-G
                    status:
                      type: string
                      enum:
                        - ACTIVE
                      description: Status of the request.
                    created_by:
                      type: integer
                      description: Identifier of the user who created the request.
                      example: 2
                    created_at:
                      type: string
                      format: date-time
                      description: Date and time when the request was created.
        '401':
          description: Unauthenticated Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Unauthenticated"
        '403':
          description: Forbidden Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Forbidden"
        404:
          description: Not Found Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Not Found"
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: Error message
                    example: "Internal Server Error"
...